+++ src/modules/opengl/shader.h
@@ -81,7 +81,7 @@ API GLuint createOpenGLShaderFromString(const char *source, GLenum type);
 API GLuint createOpenGLShaderFromFile(char *filename, GLenum type);
 API GLuint createOpenGLShaderProgram(GLuint vertexShader, GLuint fragmentShader, bool recycleShaders);
 API OpenGLUniform *createOpenGLUniformInt(int value);
-API OpenGLUniform *createOpenGLUniformFloat(float value);
+API OpenGLUniform *createOpenGLUniformFloat(double value);
 API OpenGLUniform *createOpenGLUniformVector(Vector *value);
 API OpenGLUniform *createOpenGLUniformMatrix(Matrix *value);
 API bool useOpenGLUniform(OpenGLUniform *uniform);
